results/rebuttal_new_baselines_llmatic_missing/run-20231111-034200_LLMatic-zero_shot_donnemartin-system-design-oop-eventplanner_0_20-runs_log/eventplanner/17/LLMatic/test_notifications.py
import pytest
from notifications import Notifications


def test_send_notification():
	notifications = Notifications()
	assert notifications.send_notification('user1', 'Test message') == 'Notification sent'
	assert notifications.get_notifications('user1') == ['Test message']


def test_set_reminder():
	notifications = Notifications()
	assert notifications.set_reminder('user1', 'Test reminder') == 'Reminder set'
	assert notifications.get_reminders('user1') == ['Test reminder']


def test_no_notifications():
	notifications = Notifications()
	assert notifications.get_notifications('user2') == 'No notifications'


def test_no_reminders():
	notifications = Notifications()
	assert notifications.get_reminders('user2') == 'No reminders'
